[Fix] Warningm [Fix] Open a file

+++ UFAFAC/Sources/DataStructure/StringsHolder.cpp
@@ -49,8 +49,8 @@ u32 StringsHolder::FindOrCreateString(const std::wstring& str)
 		return slot;
 	}
 	strings.push_back(str);
-	unorderedStrings.insert(std::pair<std::wstring, u32>(str, strings.size() - 1));
-	return strings.size() - 1;
+	unorderedStrings.insert(std::pair<std::wstring, u32>(str, (u32)strings.size() - (u32)1));
+	return (u32)strings.size() - (u32)1;
 }
 
 u32 StringsHolder::FindString(const std::wstring& str)
