+++ UFAFAC/Sources/DataStructure/TagManager.cpp
@@ -102,8 +102,8 @@ u32 DataStructure::TagManager::FindOrCreateTagList(const std::vector<u32>& list)
 		return slot;
 	}
 	tagsInUse.push_back(list);
-	unorderedTags.insert(std::pair<std::vector<u32>, u32>(list, tagsInUse.size() - 1));
-	return tagsInUse.size() - 1;
+	unorderedTags.insert(std::pair<std::vector<u32>, u32>(list, (u32)tagsInUse.size() - (u32)1));
+	return (u32)tagsInUse.size() - (u32)1;
 }
 
 u32 DataStructure::TagManager::FindTagList(const std::vector<u32>& list)
